feat(AbstractJobsProvider): Allow specification of providerURI in constructor
Add support for specifying the `providerURI` in the constructor of
`AbstractJobsProvider`, with a default value of `null`. This change
simplifies extending the class and offers a more straightforward way
of setting the `providerURI` when needed.

By providing a default value, the constructor remains flexible, while
making it easier to customize the `providerURI` without requiring
additional boilerplate code.